Add continue-on-error to the workflow
TruffleRuby has been failing on random benchmarks in the last three
builds:
https://github.com/ruby/ruby-bench/actions/runs/17961477542/job/51085511584
https://github.com/ruby/ruby-bench/actions/runs/17961695913/job/51086136742
https://github.com/ruby/ruby-bench/actions/runs/17967402033/job/51102476785

This job has so far failed when TruffleRuby had a bug, not when
ruby-bench introduced an incompatibility with TruffleRuby. This is not a
repository for testing TruffleRuby, so testing benchmarks with CRuby seems
enough for most of the time.

You can still see the history of GitHub Actions in case some change in
ruby-bench permanently breaks the build for TruffleRuby. I think it's
enough for maintaining the TruffleRuby support. We often don't write
CRuby-specific code other than YJIT/ZJIT integration anyway.