Fix issues with require "bundler/setup"
If your ruby's bundler is different from the saved result in
`Gemfile.lock`, bundler autoswitches by starting a new process. When it
does this it sometimes doesn't pick up the old LOAD_PATH and results in
a LoadError.

Switch to use `Bundler.setup`, which is a simpler process and uses the
current bundler.