Use an Array of floats for the json float benchmark
JSON dictionary keys are always strings, so this benchmark could be seen
to be misleading, as only the values are floats, the keys will be
implicitly converted when the hash is converted to json.

I've chosen to do use an array here instead of explicilty converting the
keys into strings because it avoids the complexity of string
interpolation, and it keeps the original intention of this micro
benchmark which was to have all the contents of the data structure
homogenous.