Add benchmarks for various Addressable URI operations (#431)
* Add benchmarks for various Addressable URI operations

We want to track the performance of common operations in the Addressable gem,
such as parsing, normalization, component access, and modification.

We heavily use the Addressable gem in our projects, so having benchmarks will help
us monitor performance changes over time and ensure that updates to the gem or our
codebase do not introduce regressions.

* Run benchmarks 10k times, not 100

---------

Co-authored-by: Max Bernstein <ruby@bernsteinbear.com>