Write a Ractor benchmark that mirrors the process structure
The previous ractor knucleotide benchmark ran the whole benchmark in
parallel inside multiple ractors.

This is a good test that the benchmark can be parallelised but isn't
directly comparable to the original Knucleotide implementation which
uses processes to partition the work done by the benchmark itself.

This implementation uses ractors to parallelise the work in the same way
that the original uses Process.fork so it's more directly comparable.

This needs to be run with the regular benchmark harness instead of the
ractor harness, otherwise the ractor harness will attempt to wrap this
benchmark run in multiple ractors too.